平均値の端数を保持する

+++ average.go
@@ -1,5 +1,5 @@
 package errorlearning
 
 func Average(total, count int) float64 {
-	return float64(total / count)
+	return float64(total) / float64(count)
 }
